# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ChoreSync API
  version: 1.0.0
  description: >-
    Frozen Contract-First OpenAPI Specification for ChoreSync - An equitable, flexible
    household chore coordination system supporting Flatmates, Families, and Couples co-living.
servers:
  - url: http://localhost:8000
    description: Local development backend server
  - url: https://api.choresync.local
    description: Production API server
...
paths:
  /api/auth/magic-link:
    post:
      tags: [Auth]
      summary: Request a magic login link or code
      description: Sends an email with a 1-click magic login link or verification code.
      operationId: requestMagicLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkRequest'
      responses:
        '200':
          description: Magic link generated or dispatched
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Magic link sent to email"
        '400':
          $ref: '#/components/responses/BadRequestError'
  /api/auth/verify:
    post:
      tags: [Auth]
      summary: Verify magic link or verification token
      description: Verifies token and returns signed JWT access token with member and household claims.
      operationId: verifyMagicLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkVerifyRequest'
      responses:
        '200':
          description: Verification successful, JWT token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/demo-login:
    post:
      tags: [Auth]
      summary: Mint JWT for demo persona fast-switch
      description: Issues a real signed JWT for seed members to enable zero-friction demo reviews.
      operationId: demoLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DemoLoginRequest'
      responses:
        '200':
          description: Demo token minted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user with email and password
      description: Verifies user credentials and returns signed JWT access token with member and household claims.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful, JWT token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register new user account
      description: Creates user account with email and password, creating a new household or joining an existing one.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created successfully, JWT token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          description: Account with email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset email
      description: Issues a password reset token and dispatches an email with reset instructions.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset email dispatched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      description: Verifies password reset token, updates password, and returns authenticated JWT session.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password successfully reset, JWT token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/households:
    get:
      tags: [Households]
      summary: List all households
      description: Returns all households accessible in the workspace.
      operationId: getHouseholds
      responses:
        '200':
          description: List of households
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Household'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Households]
      summary: Create a new household
      description: Creates a new household workspace with specified initial settings.
      operationId: createHousehold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdCreateRequest'
      responses:
        '201':
          description: Household created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /api/households/join:
    post:
      tags: [Households]
      summary: Join an existing household
      description: Joins a household using a unique invite code.
      operationId: joinHousehold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdJoinRequest'
      responses:
        '200':
          description: Joined household successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/households/{household_id}:
    get:
      tags: [Households]
      summary: Get household details
      description: Retrieves a household by its unique identifier.
      operationId: getHousehold
      parameters:
        - $ref: '#/components/parameters/HouseholdIdParam'
      responses:
        '200':
          description: Household details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/households/{household_id}/settings:
    patch:
      tags: [Households]
      summary: Update household settings
      description: Updates household configuration preferences (default mode, swap permissions, timezone).
      operationId: updateHouseholdSettings
      parameters:
        - $ref: '#/components/parameters/HouseholdIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdSettingsUpdate'
      responses:
        '200':
          description: Updated household details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/households/{household_id}/verify-pin:
    post:
      tags: [Households]
      summary: Verify household admin PIN
      description: Verifies 4-digit PIN for parent/admin gate on shared devices.
      operationId: verifyHouseholdPIN
      parameters:
        - $ref: '#/components/parameters/HouseholdIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyPINRequest'
      responses:
        '200':
          description: PIN verification successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Admin PIN verified"
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/households/{household_id}/members:
    get:
      tags: [Members]
      summary: List household members
      description: Returns all members belonging to the specified household.
      operationId: getMembers
      parameters:
        - $ref: '#/components/parameters/HouseholdIdParam'
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      tags: [Members]
      summary: Add member to household
      description: Adds or invites a new member to the household.
      operationId: addMember
      parameters:
        - $ref: '#/components/parameters/HouseholdIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberCreateRequest'
      responses:
        '201':
          description: Member added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/households/{household_id}/chores:
    get:
      tags: [Chores]
      summary: List household chores
      description: Queries chores for a household with optional filtering by status, category, view, or search query.
      operationId: getChores
      parameters:
        - $ref: '#/components/parameters/HouseholdIdParam'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ChoreStatus'
          description: Filter chores by status
        - name: category
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ChoreCategory'
          description: Filter chores by category
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search term across chore title and description
        - name: view
          in: query
          required: false
          schema:
            type: string
            enum: [all, my, open, approval, overdue]
          description: Preset view filter
        - name: assignee_id
          in: query
          required: false
          schema:
            type: string
          description: Filter by assignee ID (used with view=my)
      responses:
        '200':
          description: Filtered list of chores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chore'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      tags: [Chores]
      summary: Create a new chore
      description: Creates a new chore with assignment strategy, recurrence rules, and verification settings.
      operationId: createChore
      parameters:
        - $ref: '#/components/parameters/HouseholdIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoreCreateRequest'
      responses:
        '201':
          description: Chore created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chore'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/chores/{chore_id}:
    get:
      tags: [Chores]
      summary: Get chore details
      description: Retrieves complete details for a specific chore by ID.
      operationId: getChore
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      responses:
        '200':
          description: Chore details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chore'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags: [Chores]
      summary: Update chore
      description: Updates fields of an existing chore definition.
      operationId: updateChore
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoreUpdateRequest'
      responses:
        '200':
          description: Updated chore details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chore'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [Chores]
      summary: Delete chore
      description: Removes a chore from the household board.
      operationId: deleteChore
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      responses:
        '200':
          description: Chore deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/chores/{chore_id}/claim:
    post:
      tags: [Chores]
      summary: Claim open pool chore
      description: Assigns an unassigned or open-pool chore to the requesting member.
      operationId: claimChore
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoreClaimRequest'
      responses:
        '200':
          description: Chore claimed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chore'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/chores/{chore_id}/complete:
    post:
      tags: [Chores]
      summary: Mark chore completed
      description: >-
        Submits completion for a chore. If requires_approval is false, marks completed immediately
        and awards points. If requires_approval is true, places completion into pending_approval status.
      operationId: completeChore
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoreCompleteRequest'
      responses:
        '200':
          description: Completion result with chore state and points awarded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoreCompleteResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/chores/{chore_id}/swap:
    post:
      tags: [Swaps]
      summary: Propose swap for chore
      description: Proposes trading an assigned chore to a designated peer or onto the household swap board.
      operationId: requestChoreSwap
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoreSwapCreateRequest'
      responses:
        '201':
          description: Swap request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoreSwapRequest'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/chores/{chore_id}/nudge:
    post:
      tags: [Activity & Social]
      summary: Send gentle reminder nudge
      description: Sends a friendly reminder notification to the current assignee or open pool.
      operationId: sendNudge
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoreNudgeRequest'
      responses:
        '200':
          description: Nudge sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoreNudgeResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/chores/{chore_id}/comments:
    get:
      tags: [Activity & Social]
      summary: Get comments for chore
      description: Retrieves all discussion comments posted on a chore.
      operationId: getComments
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      responses:
        '200':
          description: List of comments with author information
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChoreCommentWithMember'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      tags: [Activity & Social]
      summary: Add comment to chore
      description: Posts a comment or question to a chore's comment feed.
      operationId: addComment
      parameters:
        - $ref: '#/components/parameters/ChoreIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoreCommentCreateRequest'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoreComment'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
  /api/uploads/photo:
    post:
      tags: [Chores]
      summary: Upload chore completion photo proof
      description: Uploads an image file to cloud object storage (S3 / Floci emulator) and returns the asset URLs.
      operationId: uploadProofPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (jpeg, png, webp, gif) up to 10MB
      responses:
        '200':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadProofPhotoResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '415':
          description: Unsupported media type
...
  /api/households/{household_id}/approvals:
    get:
      tags: [Approvals]
      summary: List pending chore approvals
      description: Returns all chore completions awaiting parent/admin verification in the household.
      operationId: getPendingApprovals
      parameters:
        - $ref: '#/components/parameters/HouseholdIdParam'
      responses:
        '200':
          description: List of pending completions with chore and member context
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PendingApprovalItem'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/completions/{completion_id}/approve:
    post:
      tags: [Approvals]
      summary: Approve chore completion
      description: Admin approves a completed chore, awarding effort points and advancing rotation/recurrence.
      operationId: approveChore
      parameters:
        - $ref: '#/components/parameters/CompletionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveCompletionRequest'
      responses:
        '200':
          description: Chore approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/completions/{completion_id}/reject:
    post:
      tags: [Approvals]
      summary: Reject chore completion
      description: Admin rejects a completion with feedback; chore reverts to assigned status for rework.
      operationId: rejectChore
      parameters:
        - $ref: '#/components/parameters/CompletionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectCompletionRequest'
      responses:
        '200':
          description: Chore completion rejected and returned for rework
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/households/{household_id}/swaps:
    get:
      tags: [Swaps]
      summary: List household chore swaps
      description: Retrieves all active and historical swap requests for a household.
      operationId: getSwaps
      parameters:
        - $ref: '#/components/parameters/HouseholdIdParam'
      responses:
        '200':
          description: List of swap requests with chore and member details
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SwapItem'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/swaps:
    post:
      tags: [Swaps]
      summary: Create swap request
      description: Proposes a new chore swap.
      operationId: createSwap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwapCreateRequest'
      responses:
        '201':
          description: Swap request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoreSwapRequest'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/swaps/{swap_id}/accept:
    post:
      tags: [Swaps]
      summary: Accept chore swap
      description: Target peer or housemate accepts a swap request; assignee responsibility transfers immediately.
      operationId: acceptSwap
      parameters:
        - $ref: '#/components/parameters/SwapIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwapAcceptRequest'
      responses:
        '200':
          description: Swap accepted and chore reassigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token encoding member_id, household_id, and role claims.
...
  responses:
    BadRequestError:
      description: Bad request or validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "BAD_REQUEST"
            message: "Title is required and cannot be empty"
            code: 400
            details: {}
    UnauthorizedError:
      description: Missing, invalid, or expired authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "UNAUTHORIZED"
            message: "Missing or invalid Bearer authentication token"
            code: 401
            details: {}
...
  schemas:
...
    MagicLinkRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "sarah@example.com"
    MagicLinkVerifyRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Magic link token or 6-digit verification code
          example: "MAGIC-123456"
...